+++ src/test/groovy/io/choerodon/devops/api/controller/v1/DevopsEnvPodControllerSpec.groovy
@@ -18,8 +18,10 @@ import org.springframework.context.annotation.Import
 import org.springframework.http.HttpEntity
 import org.springframework.http.HttpHeaders
 import org.springframework.http.MediaType
+import spock.lang.Shared
 import spock.lang.Specification
 import spock.lang.Stepwise
+import spock.lang.Subject
 
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT
 
@@ -32,10 +34,9 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
 
 @SpringBootTest(webEnvironment = RANDOM_PORT)
 @Import(IntegrationTestConfiguration)
+@Subject(DevopsEnvPodController)
 @Stepwise
 class DevopsEnvPodControllerSpec extends Specification {
-
-    private static flag = 0
 
     @Autowired
     private TestRestTemplate restTemplate
@@ -56,75 +57,76 @@ class DevopsEnvPodControllerSpec extends Specification {
     @Qualifier("mockEnvUtil")
     private EnvUtil envUtil
 
-    def setup() {
-        if (flag == 0) {
-            DevopsEnvPodDO devopsEnvPodDO = new DevopsEnvPodDO()
-            devopsEnvPodDO.setId(1L)
-            devopsEnvPodDO.setEnvId(1L)
-            devopsEnvPodDO.setAppInstanceId(1L)
-            devopsEnvPodMapper.insert(devopsEnvPodDO)
+    @Shared
+    DevopsEnvPodDO devopsEnvPodDO = new DevopsEnvPodDO()
+    @Shared
+    ApplicationVersionDO applicationVersionDO = new ApplicationVersionDO()
+    @Shared
+    ApplicationVersionDO applicationVersionDO1 = new ApplicationVersionDO()
+    @Shared
+    ApplicationDO applicationDO = new ApplicationDO()
+    @Shared
+    ApplicationDO applicationDO1 = new ApplicationDO()
+    @Shared
+    DevopsAppMarketDO devopsAppMarketDO = new DevopsAppMarketDO()
 
-            ApplicationVersionDO applicationVersionDO = new ApplicationVersionDO()
-            applicationVersionDO.setId(1L)
-            applicationVersionDO.setAppId(1L)
-            applicationVersionDO.setValueId(1L)
-            applicationVersionDO.setIsPublish(1L)
-            applicationVersionDO.setCommit("commit")
-            applicationVersionDO.setVersion("versions")
-            applicationVersionDO.setCreationDate(new Date(2018, 9, 14, 13, 40, 0))
-            applicationVersionMapper.insert(applicationVersionDO)
-            ApplicationVersionDO applicationVersionDO1 = new ApplicationVersionDO()
-            applicationVersionDO1.setId(2L)
-            applicationVersionDO1.setAppId(2L)
-            applicationVersionDO1.setValueId(1L)
-            applicationVersionDO1.setIsPublish(1L)
-            applicationVersionDO1.setCommit("commit1")
-            applicationVersionDO1.setVersion("version1")
-            applicationVersionDO1.setCreationDate(new Date(2018, 9, 14, 13, 40, 0))
-            applicationVersionMapper.insert(applicationVersionDO1)
+    def setupSpec() {
+        devopsEnvPodDO.setId(1L)
+        devopsEnvPodDO.setEnvId(1L)
+        devopsEnvPodDO.setAppInstanceId(1L)
 
-            ApplicationDO applicationDO = new ApplicationDO()
-            ApplicationDO applicationDO1 = new ApplicationDO()
-            applicationDO.setId(1L)
-            applicationDO1.setId(2L)
-            applicationDO.setActive(true)
-            applicationDO1.setActive(true)
-            applicationDO.setSynchro(true)
-            applicationDO1.setSynchro(true)
-            applicationDO.setCode("app")
-            applicationDO1.setCode("app1")
-            applicationDO.setProjectId(1L)
-            applicationDO1.setProjectId(1L)
-            applicationDO.setName("appname")
-            applicationDO1.setName("appname1")
-            applicationDO.setGitlabProjectId(1)
-            applicationDO1.setGitlabProjectId(1)
-            applicationDO.setAppTemplateId(1L)
-            applicationDO1.setAppTemplateId(1L)
-            applicationDO.setObjectVersionNumber(1L)
-            applicationDO1.setObjectVersionNumber(1L)
-            applicationMapper.insert(applicationDO)
-            applicationMapper.insert(applicationDO1)
+        applicationVersionDO.setId(1L)
+        applicationVersionDO.setAppId(1L)
+        applicationVersionDO.setValueId(1L)
+        applicationVersionDO.setIsPublish(1L)
+        applicationVersionDO.setCommit("commit")
+        applicationVersionDO.setVersion("versions")
+        applicationVersionDO.setCreationDate(new Date(2018, 9, 14, 13, 40, 0))
 
-            DevopsAppMarketDO devopsAppMarketDO = new DevopsAppMarketDO()
-            devopsAppMarketDO.setId(1L)
-            devopsAppMarketDO.setAppId(1L)
-            devopsAppMarketDO.setPublishLevel("organization")
-            devopsAppMarketDO.setContributor("testman")
-            devopsAppMarketDO.setDescription("I Love Test")
-            applicationMarketMapper.insert(devopsAppMarketDO)
+        applicationVersionDO1.setId(2L)
+        applicationVersionDO1.setAppId(2L)
+        applicationVersionDO1.setValueId(1L)
+        applicationVersionDO1.setIsPublish(1L)
+        applicationVersionDO1.setCommit("commit1")
+        applicationVersionDO1.setVersion("version1")
+        applicationVersionDO1.setCreationDate(new Date(2018, 9, 14, 13, 40, 0))
 
-//            // 创建测试用的gitops目录
-//            FileUtil.copyFile("src/test/gitops/org/pro/env/test-ing.yaml", "gitops/org/pro/env")
-//            FileUtil.copyFile("src/test/gitops/org/pro/env/test-rel.yaml", "gitops/org/pro/env")
-//            FileUtil.copyFile("src/test/gitops/org/pro/env/test-svc.yaml", "gitops/org/pro/env")
+        applicationDO.setId(1L)
+        applicationDO1.setId(2L)
+        applicationDO.setActive(true)
+        applicationDO1.setActive(true)
+        applicationDO.setSynchro(true)
+        applicationDO1.setSynchro(true)
+        applicationDO.setCode("app")
+        applicationDO1.setCode("app1")
+        applicationDO.setProjectId(1L)
+        applicationDO1.setProjectId(1L)
+        applicationDO.setName("appname")
+        applicationDO1.setName("appname1")
+        applicationDO.setGitlabProjectId(1)
+        applicationDO1.setGitlabProjectId(1)
+        applicationDO.setAppTemplateId(1L)
+        applicationDO1.setAppTemplateId(1L)
+        applicationDO.setObjectVersionNumber(1L)
+        applicationDO1.setObjectVersionNumber(1L)
 
-            flag = 1
-        }
+        devopsAppMarketDO.setId(1L)
+        devopsAppMarketDO.setAppId(1L)
+        devopsAppMarketDO.setPublishLevel("organization")
+        devopsAppMarketDO.setContributor("testman")
+        devopsAppMarketDO.setDescription("I Love Test")
     }
 
     def "PageByOptions"() {
-        given:
+        given: '初始化变量'
+        devopsEnvPodMapper.insert(devopsEnvPodDO)
+        applicationVersionMapper.insert(applicationVersionDO)
+        applicationVersionMapper.insert(applicationVersionDO1)
+        applicationMapper.insert(applicationDO)
+        applicationMapper.insert(applicationDO1)
+        applicationMarketMapper.insert(devopsAppMarketDO)
+
+        and: '设置请求头'
         String infra = null
         PageRequest pageRequest = new PageRequest(1, 20)
 
@@ -139,10 +141,10 @@ class DevopsEnvPodControllerSpec extends Specification {
         envUtil.getConnectedEnvList(_ as EnvListener) >> connectedEnvList
         envUtil.getUpdatedEnvList(_ as EnvListener) >> updateEnvList
 
-        when:
+        when: '分页查询容器管理'
         def page = restTemplate.postForObject("/v1/projects/1/app_pod/list_by_options?envId=1&appId=1", strEntity, Page.class)
 
-        then:
+        then: '校验返回值和清除数据'
         page != null
         devopsEnvPodMapper.deleteByPrimaryKey(1L)
         applicationVersionMapper.deleteByPrimaryKey(1L)
@@ -151,5 +153,4 @@ class DevopsEnvPodControllerSpec extends Specification {
         applicationMapper.deleteByPrimaryKey(2L)
         applicationMarketMapper.deleteByPrimaryKey(1L)
     }
-
 }
